chore: move file to page/log
